fix: change sql join type to inner and update corresponding table React component types
The left join meant that all the publications were retrieved regardless of whether they have successful one-time payments or not. This means that some payment information could have been null. Having the join to inner ensures that all valid payments are retrieved and only for publications that have them. Not my best moment, letting this through.